Made updates to both object detection and tracking code
Added ability to write out video with boxes and tracks.

Changed the object detection code to remove the accumulator and pass the
color image in to the MOG subtractor (instead of black and white). Also
some other misc. changes.

Changed the tracker to use both distance from last point and histogram
comparison in the cost function.  Also fixed a few bugs.